refactor profile functions into its own page
