+++ tests/e2e/tests/non_agent_flow_coolstore.test.ts
@@ -7,6 +7,7 @@ import { OPENAI_GPT4O_PROVIDER } from '../fixtures/provider-configs.fixture';
 import { KAIViews } from '../enums/views.enum';
 import { verifyAnalysisViewCleanState } from '../utilities/utils';
 import * as VSCodeFactory from '../utilities/vscode.factory';
+import { ProfilePage } from '../pages/profile.page';
 
 // NOTE: This is the list of providers that have cached data for the coolstore app
 // Update this list when you create cache for a new provider, you probably don't need
@@ -22,6 +23,7 @@ providers.forEach((config) => {
     { tag: ['@tier3', '@offline'] },
     () => {
       let vscodeApp: VSCode;
+      let profilePage: ProfilePage;
       test.beforeAll(async ({ testRepoData }: { testRepoData: any }, testInfo: any) => {
         test.setTimeout(1600000);
         const repoName = getRepoName(testInfo);
@@ -30,12 +32,13 @@ providers.forEach((config) => {
         // prepareOffline=true extracts LLM cache and sets demoMode/cacheDir BEFORE VS Code launches
         // This ensures the extension can use cached healthcheck data during initial activation
         vscodeApp = await VSCodeFactory.open(repoInfo, true, true);
+        profilePage = new ProfilePage(vscodeApp);
         try {
-          await vscodeApp.deleteProfile(profileName);
+          await profilePage.delete(profileName);
         } catch {
           console.log(`An existing profile probably doesn't exist, creating a new one`);
         }
-        await vscodeApp.createProfile(repoInfo.sources, repoInfo.targets, profileName);
+        await profilePage.create(repoInfo.sources, repoInfo.targets, profileName);
 
         await vscodeApp.configureGenerativeAI(config.config);
         await vscodeApp.startServer();
